feat: load product first and change data type

+++ api/models/product.go
@@ -3,9 +3,9 @@ package models
 type Product struct {
 	ID         int
 	Name       string
-	AVGPrice   int
-	BestPrice  int
-	GoodPrice  int
+	AVGPrice   float32
+	BestPrice  float32
+	GoodPrice  float32
 	FirstBest  []string
 	SecondBest []string
 }
